crawler: optimize main loop by iterating only over active crawlers
The crawler thread can waste a lot of CPU cycles iterating over the
inactive crawlers in this part of its main loop:

    for (i = POWER_SMALLEST; i < LARGEST_ID; i++) {
        if (crawlers[i].it_flags != 1) {
        continue;
        }
        ...
    }

In my case motivating the fix the crawler thread nearly saturates one
CPU for some time when it kicks in and causes request timeouts. Upon
investigating with "perf report" and annotating the assembly, 90% of CPU
time of the crawler thread is spent in the part of the for loop pictured
above (one slab happens to need the crawl much more frequently than
others).

To fix this, we maintain an array of sids that are currently crawled, so
that each loop iteration has a chance to do real work.